v0.2.4d: tweaked bundle getFloat to return 0.0 instead of NaN if no value is found
reasoning: we can use exists if we must check if a value is bundled, but 0.0 is a logical default in most cases. This allows code that's happy to accept 0.0 as a default to just grab the value, otherwise it would have to check for and handle NaN if there was a case where the value was not bundled.